requirements.txt
Django>=4.2,<6.0
Pillow>=10.0
qrcode>=7.4
djangorestframework>=3.15
